# feat: Optimize phrase lookup performance with pattern caching